fix(approval,vfs,kernel): CI failures + review-flagged hot-path contention
- Doctests in astrid-capabilities/src/{lib,prelude}.rs updated to the
  v2 CapabilityToken::create signature (principal) and principal-scoped
  has_capability. Caught by CI on macOS; unchanged on production.
- Extract interceptor/mod.rs test module into mod_tests.rs to stay
  under the 1000-line CI threshold (main file 443 lines).
- OverlayVfsRegistry: cache-hit path now takes a read lock and updates
  last_used via AtomicU64, so concurrent invocations for different
  principals no longer serialise on a shared write lock. Eviction
  still runs under the write lock with a relaxed snapshot.
- Kernel::connection_closed: remove the per-principal DashMap entry
  when the counter hits zero so unique-principal churn does not leak
  memory or inflate total_connection_count() scans.

All lib tests + doctests + targeted integration tests pass locally;
cargo clippy --workspace --all-features -D warnings green;
cargo fmt --all --check green.
